refactor: simplify the public package entry
Constraint: Preserve package version, public API, zero runtime dependencies, and Zustand >=4 <6 peer range
Confidence: high
Scope-risk: narrow
