--- packages/zusound/src/middleware.ts
@@ -1,4 +1,0 @@
-import { createZusound } from './adapter'
-export const zusound = createZusound()
-
-export default zusound
